pyproject.toml
[tool.poetry]
name = "real-estate-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "real_estate_service"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
sqlalchemy = "^2.0.21"
python-dotenv = "^1.0.0"
black = "^23.9.1"
flask-migrate = "^4.0.5"
flask-bootstrap = "^3.3.7.1"
flask-wtf = "^1.2.1"
flask-admin = "^1.6.0"
pytest = "^7.4.2"
pytest-env = "^1.0.1"
mixer = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 77
